Improve CMakeLists.txt and build.ps1

+++ tests/icmp/main.cpp
@@ -256,7 +256,7 @@ int main(int argc, char** argv)
   const char* host    = argc > 1 ? argv[1] : ICMPTEST_PIN_HOST;
   const int max_times = argc > 2 ? atoi(argv[2]) : 4;
 
-  printf("sizeof(intptr_t)=%zu, sizeof(wchar_t)=%zu, sizeof(long)=%zu\n", sizeof(intptr_t), sizeof(wchar_t), sizeof(long));
+  printf("sizeof(intptr_t)=%u, sizeof(wchar_t)=%u, sizeof(long)=%u\n", static_cast<int>(sizeof(intptr_t)), static_cast<int>(sizeof(wchar_t)), static_cast<int>(sizeof(long)));
 
   std::vector<ip::endpoint> endpoints;
   xxsocket::resolve(endpoints, host, 0);
